moving steps tests to /tests folder
