+++ src/step-by-step/examples/captcha/recipe.json
