+++ src/step-by-step/examples/cnes/recipe.json
